resize client

+++ src/Services/Web/Business/Models/ResizeImageRequest.cs
@@ -4,8 +4,8 @@ public class ResizeImageRequest
 {
     public byte[] Image { get; set; }
     public string? FileName { get; set; }
-    public decimal ResizeWidth { get; set; }
-    public decimal ResizeHeight { get; set; }
+    public double ResizeWidth { get; set; }
+    public double ResizeHeight { get; set; }
     public long ImageSize {get ;set ;}
     public int OriginalWidth { get; set; }
     public int OriginalHeight { get; set; }
